fix(lua): rawlen in mingw32

+++ ffi/lua/src/lua.c
@@ -174,7 +174,7 @@ void set_name(lua_State *L, int idx, const char *name, const char *data)
 
 void arraynum_get(lua_State *L, int idx, lua_Number *ptr, unsigned int num)
 {
-    size_t len = lua_rawlen(L, idx);
+    size_t len = (size_t)lua_rawlen(L, idx);
     if (num > len)
     {
         num = (unsigned int)len;
@@ -232,7 +232,7 @@ void arraynum_sets(lua_State *L, int idx, const SFnums *tab)
 size_t tablenum_len(lua_State *L, int idx)
 {
     size_t num = 0;
-    size_t n = lua_rawlen(L, idx);
+    size_t n = (size_t)lua_rawlen(L, idx);
     for (unsigned int i = 0; i++ != n;)
     {
         lua_rawgeti(L, idx, (int)i);
@@ -252,7 +252,7 @@ size_t tablenum_len(lua_State *L, int idx)
 
 lua_Number *tablenum_num(lua_State *L, int idx, lua_Number *ptr)
 {
-    size_t n = lua_rawlen(L, idx);
+    size_t n = (size_t)lua_rawlen(L, idx);
     for (unsigned int i = 0; i++ != n;)
     {
         lua_rawgeti(L, idx, (int)i);
